minor: bug fix

+++ src/Feedback/User_Template/FB_User_Template.cpp
@@ -76,7 +76,7 @@ extern void (*FB_End_User_Ptr)();
 // Return      :  Fluid, ParAtt
 //-------------------------------------------------------------------------------------------------------
 int FB_User_Template( const int lv, const double TimeNew, const double TimeOld, const double dt,
-                      const int NPar, const int *ParSortID, real *ParAtt[PAR_NATT_TOTAL],
+                      const int NPar, const long *ParSortID, real *ParAtt[PAR_NATT_TOTAL],
                       real (*Fluid)[FB_NXT][FB_NXT][FB_NXT], const double EdgeL[], const double dh, bool CoarseFine[],
                       const int TID, RandomNumber_t *RNG )
 {
